♻️ [refactor] 환율데이터 연결 및 원화 소수점 자리 제거
1. 날짜데이터 넣어서 환율 입력
2. 원화 소수점 자리 제거

+++ TripLog/TripLog/Feat-Hamsik/Extension/MyCashBookEntity+Ext.swift
@@ -12,7 +12,7 @@ import CoreData
 struct MockMyCashBookModel {
     let amount: Double // 지출금액
     let cashBookID: UUID // 가계부 Entity ID
-    let caculatedAmount: Int // 계산값(원화)
+    let caculatedAmount: Double // 계산값(원화)
     let category: String // 카테고리
     let country: String // 환율코드
     let expenseDate: Date // 지출일자
